move .vscode to user-level preferences
I should probably point to a style guide; using gofmt, ESLint, and Prettier defaults for most formatting.. but the more important part of the style guide is choice of language features and overall file organization.

Punting on details until there are actually multiple contributors to these projects.  The demo/MVP can be a solo project at early stages, it's fine